refactor: extract impl Config to lifecycle module (confgi-1) (#2874)
Move main.rs into a directory module and pull the `impl Config` block
into a sibling main/lifecycle.rs
